test(#3515): split the awaited-stream tests out (campsite rule)
`harness_tests.rs` reached 1510 lines with the four round-18 tests in it — 10
past the 1500-line test threshold — so they move to a child module, by
SUBJECT, following `budgets.rs`/`budgets/census_tests.rs`. A child module
sees its parent's private items, so the tests keep using the real
`ChildIo::synthetic`/`mark_from_the_start` helpers.
